requirements.txt
SpeechRecognition==3.14.3
python-dotenv==1.0.1
PyYAML==6.0.2
PyAudio==0.2.14